# DPL: Cleanup and optimization of DataRelayer benchmark